requirements.txt
APScheduler==3.10.4
aiogram==2.24
aiohttp==3.8.5
asyncpg==0.27.0
bs4==0.0.1
python-dotenv==1.0.0
urllib3==2.0.5